コミュニティ画面は非ログインユーザーも観れるように修正
