+++ app/Http/Controllers/MatchingController.php
@@ -88,76 +88,76 @@ class MatchingController extends Controller
         return redirect()->route('community',['community_id' => $community_id]);
     }
 
-    public function matching_community()
-    {
-        $communitys = Community::query()
-            ->select('id','community_name','community_image','community_members','created_at','updated_at')
-            ->orderBy('community_members', 'desc')
-            ->simplePaginate(12);
+    // public function matching_community()
+    // {
+    //     $communitys = Community::query()
+    //         ->select('id','community_name','community_image','community_members','created_at','updated_at')
+    //         ->orderBy('community_members', 'desc')
+    //         ->simplePaginate(12);
+    //
+    //     $my_community = \DB::table('user_community')
+    //         ->select('community_id')
+    //         ->where('user_id',Auth::user()->id)
+    //         ->get();
+    //
+    //     $my_communitys = [];
+    //     foreach($my_community as $id){
+    //         array_push($my_communitys,$id->community_id);
+    //     }
+    //
+    //     return view('matching.matching_community',compact('communitys','my_communitys'));
+    // }
 
-        $my_community = \DB::table('user_community')
-            ->select('community_id')
-            ->where('user_id',Auth::user()->id)
-            ->get();
-
-        $my_communitys = [];
-        foreach($my_community as $id){
-            array_push($my_communitys,$id->community_id);
-        }
-
-        return view('matching.matching_community',compact('communitys','my_communitys'));
-    }
-
-    public function search_community(Request $request)
-    {
-        $select;
-        $order;
-        switch ($request->select) {
-            case 1:
-                $select = 'community_members';
-                $order = 'desc';
-                break;
-            case 2:
-                $select = 'community_members';
-                $order = 'asc';
-                break;
-            case 3:
-                $select = 'created_at';
-                $order = 'desc';
-                break;
-            case 4:
-                $select = 'created_at';
-                $order = 'asc';
-                break;
-            case 5:
-                $select = 'updated_at';
-                $order = 'desc';
-                break;
-            case 6:
-                $select = 'updated_at';
-                $order = 'asc';
-                break;
-        }
-
-        //検索
-        $communitys = Community::query()
-            ->select('id','community_name','community_image','community_members','created_at','updated_at')
-            ->where('community_name', 'LIKE', "%{$request->community_name}%")
-            ->orderBy($select, $order)
-            ->simplePaginate(100);
-
-        $my_community = \DB::table('user_community')
-            ->select('community_id')
-            ->where('user_id',Auth::user()->id)
-            ->get();
-
-        $my_communitys = [];
-        foreach($my_community as $id){
-            array_push($my_communitys,$id->community_id);
-        }
-
-        return view('matching.search_community',compact('communitys','my_communitys'));
-    }
+    // public function search_community(Request $request)
+    // {
+    //     $select;
+    //     $order;
+    //     switch ($request->select) {
+    //         case 1:
+    //             $select = 'community_members';
+    //             $order = 'desc';
+    //             break;
+    //         case 2:
+    //             $select = 'community_members';
+    //             $order = 'asc';
+    //             break;
+    //         case 3:
+    //             $select = 'created_at';
+    //             $order = 'desc';
+    //             break;
+    //         case 4:
+    //             $select = 'created_at';
+    //             $order = 'asc';
+    //             break;
+    //         case 5:
+    //             $select = 'updated_at';
+    //             $order = 'desc';
+    //             break;
+    //         case 6:
+    //             $select = 'updated_at';
+    //             $order = 'asc';
+    //             break;
+    //     }
+    //
+    //     //検索
+    //     $communitys = Community::query()
+    //         ->select('id','community_name','community_image','community_members','created_at','updated_at')
+    //         ->where('community_name', 'LIKE', "%{$request->community_name}%")
+    //         ->orderBy($select, $order)
+    //         ->simplePaginate(100);
+    //
+    //     $my_community = \DB::table('user_community')
+    //         ->select('community_id')
+    //         ->where('user_id',Auth::user()->id)
+    //         ->get();
+    //
+    //     $my_communitys = [];
+    //     foreach($my_community as $id){
+    //         array_push($my_communitys,$id->community_id);
+    //     }
+    //
+    //     return view('matching.search_community',compact('communitys','my_communitys'));
+    // }
 
     public function verify_community($community_id)
     {
